requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Django==3.2.5
django-crispy-forms==1.12.0
djangorestframework==3.12.4
future==0.18.2
html5lib==1.1
idna==3.2
joblib==1.0.1
MeaningCloud-python==2.0.0
nltk==3.6.2
Pillow==8.3.1
pycodestyle==2.7.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
regex==2021.7.6
reportlab==3.5.68
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tqdm==4.61.2
urllib3==1.26.6
webencodings==0.5.1
xhtml2pdf==0.2.5
